521 Add preview chart (#551)
* change trajectory data structure to fit graph

* move spark-chart to elements

* fix spark-chart path

* fix launch preview styling

* add spark-chart to launch preview

* cleanup

* refactor preview- working

* apply requested changes
