+++ src/lbpDashboard/lbpDashboard.html
@@ -2,7 +2,7 @@
   <require from="./projectTokenInfo/projectTokenInfo"></require>
   <require from="./lbpDashboardForm/lbpDashboardForm"></require>
   <require from="./tokenDistribution/tokenDistribution"></require>
-  <require from="./sparkChart/spark-chart"></require>
+  <require from="resources/elements/sparkChart/spark-chart"></require>
   <require from="./timeRemaining/timeRemaining"></require>
   <require from="./lbpDetails/lbpDetails"></require>
   <require from="./statistics/statistics"></require>
