--- src/lbpDashboard/sparkChart/spark-chart.scss
@@ -1,4 +1,0 @@
-.sparkChartContainer {
-  // position: relative;
-  height: 100%;
-}
