+++ src/resources/elements/sparkChart/spark-chart.html
