+++ src/resources/elements/sparkChart/spark-chart.scss
@@ -1,0 +1,8 @@
+@import "styles/styles.scss";
+.sparkChartContainer {
+  // position: relative;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
