+++ src/services/ProjectTokenHistoricalPriceService.ts
@@ -41,20 +41,25 @@ export class ProjectTokenHistoricalPriceService {
   }
 
   /**
-   * get project token price history, in USD
-   * @param lbpMgr LbpManager
+   * Get Project Token Price History, in USD
+   *
+   * The timepoints in the returned array will be relative to the UTC timezone.
+   * startTime and endTime, which, according to the LBPManager entity, are in the user's timezone would be converted to UTC.
+   *
+   * @param lbpMgr LBP Manager object
+   * @returns Array(IHistoricalPriceRecord): {time: number, price?: number}
    */
   public async getPricesHistory(lbpMgr: LbpManager): Promise<Array<IHistoricalPriceRecord>> {
     if (!lbpMgr.lbp || !lbpMgr.lbp.poolId) {
       return [];
     }
 
-    const startingSeconds = lbpMgr.startTime.getTime() / 1000;
+    const startingSeconds = this.dateService.translateLocalToUtc(lbpMgr.startTime).getTime() / 1000;
     const intervalMinutes = 60/*min*/;
     const intervalSeconds = intervalMinutes * 60/* sec */;
     const startTime = (Math.floor(startingSeconds / intervalSeconds) * intervalSeconds)/* Rounded */;
     /* Rounded to the nearest hour */
-    const endTimeSeconds = Math.floor(new Date().getTime() / 1000 / intervalSeconds) * intervalSeconds + intervalSeconds; // rounded hour
+    const endTimeSeconds = Math.floor(this.dateService.translateLocalToUtc(new Date()).getTime() / 1000 / intervalSeconds) * intervalSeconds + intervalSeconds; // rounded hour
 
 
     /**
@@ -132,13 +137,11 @@ export class ProjectTokenHistoricalPriceService {
           }
         }
 
-        const timezoneOffset = new Date().getTimezoneOffset() * 60;
         const priceAtTimePoint = fundingTokenPricesUSD.filter(price => price.timestamp <= timestamp );
-
 
         if (todaysSwaps?.length) {
           returnArray.push({
-            time: timestamp - timezoneOffset + intervalSeconds/* Apply to the next interval in users timezone */,
+            time: timestamp + intervalSeconds/* Apply to the next interval in users timezone */,
             price: (
               this.numberService.fromString(todaysSwaps[todaysSwaps.length-1].tokenAmountIn) /
               this.numberService.fromString(todaysSwaps[todaysSwaps.length-1].tokenAmountOut) *
@@ -154,7 +157,7 @@ export class ProjectTokenHistoricalPriceService {
            * previous value effected by USD course change
            */
           returnArray.push({
-            time: timestamp - timezoneOffset + intervalSeconds/* Apply to the next interval in users timezone */,
+            time: timestamp + intervalSeconds/* Apply to the next interval in users timezone */,
             price: (
               previousTimePoint *
               priceAtTimePoint[priceAtTimePoint.length-1].priceInUSD
@@ -162,7 +165,7 @@ export class ProjectTokenHistoricalPriceService {
           });
         } else {
           returnArray.push({
-            time: timestamp - timezoneOffset + intervalSeconds/* Apply to the next interval in users timezone */,
+            time: timestamp + intervalSeconds/* Apply to the next interval in users timezone */,
           });
         }
       }
